Fix errors, bugs and Add test for Car, Bicycle, Moped classes

+++ kosowski/workshop/library/src/model/Moped.cpp
@@ -1,4 +1,4 @@
 #include "model/Moped.h"
 #include <cstring>
 
-Moped::Moped(const std::string &mopPNumber, const unsigned &mopBPrice, const int &mopEDisplac): MotorVehivle(mopPNumber, mopBPrice, mopEDisplac) {}
+Moped::Moped(const std::string &mopPNumber, const double &mopBPrice, const int &mopEDisplac): MotorVehivle(mopPNumber, mopBPrice, mopEDisplac) {}
